requirements.txt
numpy
Pillow
flask
python-magic
python-magic-bin
starlette
fastapi